requirements.txt
streamlit
opencv-python~=4.11.0.86
numpy~=2.2.3
face-recognition~=1.3.0
pillow~=11.1.0